+++ test/unit/mongo_rep_admin/process_json.py
@@ -42,9 +42,9 @@ class Mail(object):
     Description:  Class stub holder for gen_class.Mail class.
 
     Methods:
-        __init__ -> Class initialization.
-        add_2_msg -> Stub method holder for Mail.add_2_msg.
-        send_mail -> Stub method holder for Mail.send_mail.
+        __init__
+        add_2_msg
+        send_mail
 
     """
 
@@ -55,7 +55,6 @@ class Mail(object):
         Description:  Class initialization.
 
         Arguments:
-            None
 
         """
 
@@ -76,17 +75,23 @@ class Mail(object):
 
         return True
 
-    def send_mail(self):
+    def send_mail(self, use_mailx=False):
 
         """Method:  get_name
 
         Description:  Stub method holder for Mail.send_mail.
 
         Arguments:
+            (input) use_mailx -> True|False - To use mailx command.
 
         """
 
-        return True
+        status = True
+
+        if use_mailx:
+            status = True
+
+        return status
 
 
 class UnitTest(unittest.TestCase):
@@ -96,16 +101,17 @@ class UnitTest(unittest.TestCase):
     Description:  Class which is a representation of a unit testing.
 
     Methods:
-        setUp -> Initialize testing environment.
-        test_mongo_failure -> Test with failed connection to Mongo.
-        test_mongo_successful -> Test with successful connection to Mongo.
-        test_json_flatten -> Test with JSON format to standard out suppression.
-        test_json_stdout_suppress -> Test with JSON format std out suppression.
-        test_json_stdout -> Test with JSON format to standard out.
-        test_mongo -> Test with writing to mongo.
-        test_file_append -> Test with writing to file in append mode.
-        test_file -> Test with writing to file.
-        test_email -> Test with email option.
+        setUp
+        test_mongo_failure
+        test_mongo_successful
+        test_json_flatten
+        test_json_stdout_suppress
+        test_json_stdout
+        test_mongo
+        test_file_append
+        test_file
+        test_email_mailx
+        test_email
 
     """
 
@@ -128,6 +134,7 @@ class UnitTest(unittest.TestCase):
                                                "%Y-%m-%d %H:%M:%S"),
             "Slaves": ["slave1", "slave2"]}
         self.args_array = {"-z": True}
+        self.args_arraya = {"-z": True, "-u": True}
         self.args_array2 = {}
         self.args_array3 = {"-z": True, "-a": True}
         self.args_array4 = {"-z": True, "-f": True}
@@ -270,6 +277,21 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mongo_rep_admin._process_json(
                 self.outdata, ofile="Filename", args_array=self.args_array),
+            self.status)
+
+    def test_email_mailx(self):
+
+        """Function:  test_email_mailx
+
+        Description:  Test with email option using mailx.
+
+        Arguments:
+
+        """
+
+        self.assertEqual(
+            mongo_rep_admin._process_json(
+                self.outdata, mail=self.mail, args_array=self.args_arraya),
             self.status)
 
     def test_email(self):